refactor: extract breadcrumb sync helpers to module
